+++ src/Ext/BuildingType/Body.cpp
@@ -283,6 +283,12 @@ void BuildingTypeExt::LoadFromINIFile(CCINIClass* const pINI)
 
 	this->RevealToAll_Radius.Read(exINI, pSection, "RevealToAll.Radius");
 
+	this->NewEvaVoice_Tag.Read(pINI, pSection, "NewEVAVoice.Tag");
+	this->NewEvaVoice_Priority.Read(exINI, pSection, "NewEVAVoice.Priority");
+	this->NewEvaVoice_RecheckOnDeath.Read(exINI, pSection, "NewEVAVoice.RecheckOnDeath");
+	this->NewEvaVoice_InitialMessage.Read(exINI, pSection, "NewEVAVoice.InitialMessage");
+	this->NewEvaVoice_EndingMessage.Read(exINI, pSection, "NewEVAVoice.EndingMessage");
+
 	if (pThis->NumberOfDocks > 0)
 	{
 		std::optional<DirType> empty;
@@ -471,6 +477,11 @@ void BuildingTypeExt::Serialize(T& Stm)
 		.Process(this->BuildingRepairedSound)
 		.Process(this->Refinery_UseNormalActiveAnim)
 		.Process(this->HasPowerUpAnim)
+		.Process(this->NewEvaVoice_Tag)
+		.Process(this->NewEvaVoice_Priority)
+		.Process(this->NewEvaVoice_RecheckOnDeath)
+		.Process(this->NewEvaVoice_InitialMessage)
+		.Process(this->NewEvaVoice_EndingMessage)
 		.Process(this->UndeploysInto_Sellable)
 		.Process(this->BuildingRadioLink_SyncOwner)
 		.Process(this->GuardRetryDelay)