+++ src/Ext/HouseType/Hooks.cpp
@@ -1,6 +1,7 @@
 #include "Body.h"
 
 #include <Ext/Scenario/Body.h>
+#include <Ext/Side/Body.h>
 
 DEFINE_HOOK(0x535005, ScenarioClass_LoadSide_SetEVAIndex, 0x6)
 {
@@ -12,12 +13,7 @@ DEFINE_HOOK(0x535005, ScenarioClass_LoadSide_SetEVAIndex, 0x6)
 	else if (SessionClass::Instance.IsCampaign())
 	{
 		if (const auto pHouse = HouseClass::CurrentPlayer)
-		{
-			const int EVAIndex = HouseTypeExt::Fetch(pHouse->Type)->EVATag;
-
-			if (EVAIndex != -2)
-				VoxClass::EVAIndex = EVAIndex;
-		}
+			VoxClass::EVAIndex = SideExt::GetOwnerEVAIndex(pHouse);
 	}
 
 	return 0;
@@ -26,12 +22,7 @@ DEFINE_HOOK(0x535005, ScenarioClass_LoadSide_SetEVAIndex, 0x6)
 DEFINE_HOOK(0x68AD0C, ScenarioClass_ReadMap_SetEVAIndex, 0x7)
 {
 	if (const auto pHouse = HouseClass::CurrentPlayer)
-	{
-		const int EVAIndex = HouseTypeExt::Fetch(pHouse->Type)->EVATag;
-
-		if (EVAIndex != -2)
-			VoxClass::EVAIndex = EVAIndex;
-	}
+		VoxClass::EVAIndex = SideExt::GetOwnerEVAIndex(pHouse);
 
 	return 0;
 }